fix: checked status of data filters options
